requirements.txt
# --- Core web framework ---
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==22.0.0

# --- Scraping and browser automation ---
playwright==1.53.0
selenium==4.37.0
beautifulsoup4==4.13.4
requests-html==0.10.0
fake-useragent==2.2.0
brotli==1.2.0
aiohttp==3.12.13

# --- AI, ML, and LangChain ---
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.3
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.3.45
sentence-transformers==5.0.0
transformers==4.53.1
scikit-learn==1.7.0
faiss-cpu==1.12.0
torch==2.7.1
pymupdf== 1.26.7


# --- Async, HTTP, and utility ---
httpx==0.28.1
httpx-sse==0.4.0
anyio==4.9.0
aiofiles==23.2.1
tenacity==8.5.0
requests==2.32.5
requests-toolbelt==1.0.0
python-dotenv==1.0.1
orjson==3.10.18
PyYAML==6.0.2


# --- Data, parsing, and misc ---
cachetools==5.5.2
numpy==2.2.6
scipy==1.16.0
pandas==2.3.0
networkx==3.5
dataclasses-json==0.6.7
MarkupSafe==3.0.2
Jinja2==3.1.6
Pygments==2.19.1
regex==2024.11.6
packaging==24.2
setuptools==80.9.0
tqdm==4.67.1
attrs==25.3.0
colorama==0.4.6
filelock==3.18.0
fonttools==4.58.5
pillow==11.2.1
multidict==6.5.0
frozenlist==1.7.0
zstandard==0.23.0
xxhash==3.5.0

# --- Pydantic and typing ---
pydantic==2.11.7
pydantic-settings==2.11.0
pydantic_core==2.33.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0

# --- Test ---
pytest==8.4.1
pytest-asyncio==1.2.0

# --- Google/Cloud ---
google-cloud-storage==3.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-genai==1.51.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0

# --- Other ---
defusedxml==0.7.1
ratelimit==2.2.1
idna==3.10
joblib==1.5.1
SQLAlchemy==2.0.41
rank_bm25==0.2.2
redis==5.0.7
